front_end/requirements_front.txt
streamlit
tensorflow
pillow
patchify
numpy
pandas